Make MaitreD public
The changes in this commit was driven by the test. I first wrote the
test and had to jump through quite a few hoops to make it compile (and
still fail when executed), but I succeeded in doing that. After that, I
passed the test with the simplest change I could think of. I wasn't
particularly applying the Devil's Advocate technique, so I went for a
decent implementation.

Considering Red Green Refactor, I haven't yet done the Refactor step. I
plan to do that now, but in separate commits so that an interested
reader can follow along.